transaction-service/requirements.txt
grpcio==1.3.0
grpcio-tools==1.3.0
redis==2.10.5
protobuf3-to-dict==0.1.5